+++ onnx/defs/math/defs.cc
@@ -3102,7 +3102,7 @@ ONNX_OPERATOR_SET_SCHEMA(
           if (axis_tensor->dims_size() != 0) {
             fail_shape_inference("axis input must be a scalar.");
           }
-          const int64_t axis = GetScalarValueFromTensor<int64_t>(axis_tensor);
+          const int64_t axis = defs::math::utils::GetScalarValueFromTensor<int64_t>(axis_tensor);
           const int64_t rank = input_shape.dim_size();
 
           if (!(-rank <= axis && axis < rank)) {
@@ -3124,7 +3124,7 @@ ONNX_OPERATOR_SET_SCHEMA(
             if (dft_length->dims_size() != 0) {
               fail_shape_inference("dft_length input must be a scalar.");
             }
-            auto dft_length_value = GetScalarValueFromTensor<int64_t>(dft_length);
+            auto dft_length_value = defs::math::utils::GetScalarValueFromTensor<int64_t>(dft_length);
             result_shape_proto.mutable_dim(axis_idx)->set_dim_value(dft_length_value);
           }
 
@@ -3208,7 +3208,7 @@ Generates a {name} window as described in the paper https://ieeexplore.ieee.org/
         fail_shape_inference("size input must be a scalar.");
       }
 
-      auto size_value = GetScalarValueFromTensor<int64_t>(size);
+      auto size_value = defs::math::utils::GetScalarValueFromTensor<int64_t>(size);
       if (size_value <= 0) {
         fail_shape_inference("size input must be greater than 0.");
       }
@@ -3438,12 +3438,12 @@ ONNX_OPERATOR_SET_SCHEMA(
           if (num_mel_bins->dims_size() != 0) {
             fail_shape_inference("num_mel_bins input must be scalar.");
           }
-          num_mel_bins_value = GetScalarValueFromTensor<int64_t>(num_mel_bins);
+          num_mel_bins_value = defs::math::utils::GetScalarValueFromTensor<int64_t>(num_mel_bins);
 
           if (dft_length->dims_size() != 0) {
             fail_shape_inference("dft_length input must be scalar.");
           }
-          dft_length_value = GetScalarValueFromTensor<int64_t>(dft_length);
+          dft_length_value = defs::math::utils::GetScalarValueFromTensor<int64_t>(dft_length);
 
           if (num_mel_bins_value > 0 && dft_length_value > 0) {
             ONNX_NAMESPACE::TensorShapeProto result_shape;
@@ -3555,7 +3555,7 @@ ONNX_OPERATOR_SET_SCHEMA(
           if (nullptr == frame_step) {
             return;
           }
-          auto frame_step_value = GetScalarValueFromTensor<int64_t>(frame_step);
+          auto frame_step_value = defs::math::utils::GetScalarValueFromTensor<int64_t>(frame_step);
 
           // Determine the size of the DFT based on the 2 optional inputs window and frame_length.
           // One must be set.
@@ -3585,7 +3585,7 @@ ONNX_OPERATOR_SET_SCHEMA(
             if (frame_length->dims_size() != 0) {
               fail_shape_inference("frame_length input must be scalar.");
             }
-            auto frame_length_value = GetScalarValueFromTensor<int64_t>(frame_length);
+            auto frame_length_value = defs::math::utils::GetScalarValueFromTensor<int64_t>(frame_length);
 
             // Ensure that the window length and the dft_length match.
             if (window_shape->dim_size() != 1) {
@@ -3616,7 +3616,7 @@ ONNX_OPERATOR_SET_SCHEMA(
             if (frame_length->dims_size() != 0) {
               fail_shape_inference("frame_length input must be scalar.");
             }
-            dft_size = GetScalarValueFromTensor<int64_t>(frame_length);
+            dft_size = defs::math::utils::GetScalarValueFromTensor<int64_t>(frame_length);
           }
 
           bool is_onesided = static_cast<bool>(getAttribute(ctx, "onesided", 0));
